+++ src/playerbuttons.cpp
@@ -22,6 +22,9 @@
 
 #include <QAbstractButton>
 #include <QMediaPlaylist>
+#ifdef Q_OS_WIN
+#include <QWinTaskbarProgress>
+#endif
 
 QMediaPlayer *PlayerButtons::currentlyPlaying = nullptr;
 
@@ -45,11 +48,18 @@ QMediaPlayer *PlayerButtons::mediaPlayer() const
 
 void PlayerButtons::setMediaPlayer(QMediaPlayer *mediaPlayer)
 {
+    if (m_mediaPlayer != nullptr) {
+        disconnect(m_mediaPlayer, &QMediaPlayer::stateChanged, this, &PlayerButtons::loadPlayerState);
+        disconnect(m_mediaPlayer, &QMediaPlayer::positionChanged, this, &PlayerButtons::processPositionChanged);
+    }
+
     m_mediaPlayer = mediaPlayer;
     if (m_mediaPlayer->playlist() == nullptr)
         m_mediaPlayer->setPlaylist(new QMediaPlaylist);
 
     connect(m_mediaPlayer, &QMediaPlayer::stateChanged, this, &PlayerButtons::loadPlayerState);
+    connect(m_mediaPlayer, &QMediaPlayer::positionChanged, this, &PlayerButtons::processPositionChanged);
+
     loadPlayerState(m_mediaPlayer->state());
 }
 
@@ -83,6 +93,7 @@ void PlayerButtons::loadPlayerState(QMediaPlayer::State state)
 
         m_playPauseButton->setIcon(QIcon::fromTheme("media-playback-start"));
         m_stopButton->setEnabled(false);
+        emit stopped();
         break;
     case QMediaPlayer::PlayingState:
         if (currentlyPlaying != nullptr)
@@ -91,12 +102,14 @@ void PlayerButtons::loadPlayerState(QMediaPlayer::State state)
 
         m_playPauseButton->setIcon(QIcon::fromTheme("media-playback-pause"));
         m_stopButton->setEnabled(true);
+        emit played();
         break;
     case QMediaPlayer::PausedState:
         if (currentlyPlaying == m_mediaPlayer)
             currentlyPlaying = nullptr;
 
         m_playPauseButton->setIcon(QIcon::fromTheme("media-playback-start"));
+        emit paused();
         break;
     }
 }
@@ -114,4 +127,10 @@ void PlayerButtons::processPlayPausePressed()
         m_mediaPlayer->play();
         break;
     }
+}
+
+void PlayerButtons::processPositionChanged(qint64 position)
+{
+    if (m_mediaPlayer->duration() != 0)
+        emit positionChanged(static_cast<int>(position * 100 / m_mediaPlayer->duration()));
 }